rgw: move RGWProcessEnv into new rgw_process_env.h
